backend/requirements.txt
asgiref==3.8.1
boto3==1.35.81
botocore==1.35.81
Django==5.1.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-crispy-forms==1.14.0
django-environ==0.11.2
django-mathfilters==1.0.0
django-storages==1.14.4
django-utils-six==2.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.2
urllib3==2.2.3
